chore: getting updateLabels out of the component
